Combined Andres code and setup is done for local api
