+++ Dockerfile
@@ -1,8 +1,7 @@
 FROM python:3.10.6-buster
 
-WORKDIR /test_aw
+WORKDIR /ratemate_app_test
 COPY requirements.txt requirements.txt
-
 
 RUN pip install --upgrade pip
 
@@ -11,6 +10,7 @@ RUN pip install -r requirements.txt
 COPY api api
 COPY ml_logic ml_logic
 COPY setup.py setup.py
+
 RUN pip install .
