--- requirements_phani.txt
@@ -1,27 +1,0 @@
-torch
-seaborn
-pathlib
-colorama
-matplotlib
-transformers
-
-google-cloud-storage
-google-cloud-bigquery
-
-numpy==1.23.5
-pandas==1.5.3
-mlflow==2.1.1
-scikit-learn==1.3.1
-
-python-dotenv
-
-fastapi
-uvicorn
-
-# pytz
-# httpx
-# pytest-asyncio
-# db-dtypes
-# pyarrow
-# prefect==2.8.1
-# psycopg2-binary
