--- templates/predict_result.html
@@ -1,16 +1,0 @@
-<!-- predict_result.html -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Prediction Result</title>
-    <link rel="stylesheet" href="{{ url_for('static', path='styles.css') }}">
-</head>
-<body>
-
-<h1>Prediction Result</h1>
-<p>{{ result.prediction }}</p>
-
-</body>
-</html>
